Move root soilgrids scripts into tools/soilgrids
